+++ btclib/ecc/dsa.py
@@ -50,7 +50,6 @@ def _serialize_scalar(scalar: int) -> bytes:
 
 
 def _deserialize_scalar(sig_data_stream: BytesIO, strict: bool = True) -> int:
-
     marker = sig_data_stream.read(1)
     if marker != _DER_SCALAR_MARKER:
         err_msg = f"invalid value header: {marker.hex()}"
@@ -165,7 +164,12 @@ class Sig:
         return _DER_SIG_MARKER + var_bytes.serialize(out)
 
     @classmethod
-    def parse(cls: type[Sig], data: BinaryData, check_validity: bool = True, strict: bool = True) -> Sig:
+    def parse(
+        cls: type[Sig],
+        data: BinaryData,
+        check_validity: bool = True,
+        strict: bool = True,
+    ) -> Sig:
         """Return a Sig by parsing binary data.
 
         Deserialize a strict ASN.1 DER representation of an ECDSA
@@ -186,8 +190,8 @@ class Sig:
 
         # [0x02][r-size][r][0x02][s-size][s]
         sig_data_substream = bytesio_from_binarydata(sig_data)
-        r = _deserialize_scalar(sig_data_substream, scalar)
-        s = _deserialize_scalar(sig_data_substream, saclar)
+        r = _deserialize_scalar(sig_data_substream, strict)
+        s = _deserialize_scalar(sig_data_substream, strict)
 
         # to prevent malleability
         # the sig_data_substream must have been consumed entirely