+++ tests/tx/test_tx.py
@@ -15,7 +15,7 @@ from os import path
 import pytest
 
 from btclib.exceptions import BTClibValueError
-from btclib.script import Witness, ScriptPubKey
+from btclib.script import ScriptPubKey, Witness
 from btclib.tx import OutPoint, Tx, TxIn, TxOut, join_txs